Separated utility functions from CLI functions
